Upgrades are now purchasable

+++ Classes/Upgrade.cs
@@ -14,7 +14,7 @@ namespace ZombieSim
         public string Name { get; set; } //name of the upgrade
         public string Picture { get; set; } //picture of the upgrade
         public string Description { get; set; } //description of the upgrade
-        public float Cost { get; set; } //cost of buying this upgrade
+        public long Cost { get; set; } //cost of buying this upgrade
         public int Count { get; set; } //amount of the particular upgrade
         public int ZombiesPerTick { get; set; } //amount of zombie per second given by this particular upgrade
         public int ZombiesPerClick { get; set; } //amount of zombie per click given by this particular upgrade
